Move getFormAuthorizationState into its own module
